Fix precision issue in calibration error (#1919)
* fix implementation

* add tests

* changelog

* skip on older versions

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* skip testing on older

---------


(cherry picked from commit 879595de67ab35d891a227e3254e0b5a26a050f0)

+++ src/torchmetrics/functional/classification/calibration_error.py
@@ -82,7 +82,7 @@ def _ce_compute(
         Tensor: Calibration error scalar.
     """
     if isinstance(bin_boundaries, int):
-        bin_boundaries = torch.linspace(0, 1, bin_boundaries + 1, dtype=torch.float, device=confidences.device)
+        bin_boundaries = torch.linspace(0, 1, bin_boundaries + 1, dtype=confidences.dtype, device=confidences.device)
 
     if norm not in {"l1", "l2", "max"}:
         raise ValueError(f"Argument `norm` is expected to be one of 'l1', 'l2', 'max' but got {norm}")
